Move svg and png route functions to operations
And rename them to indicate maybe create.
